requirements.txt
django
djangorestframework
langchain
langchain-google-genai
python-dotenv
uvicorn
pypdf
tiktoken
google-generativeai
pinecone
celery
redis
gevent
eventlet